Scope Api's Project logic inside Neighborly::Api module
Following this approach, we avoid conflicts between engines and keep the
"core" model (Project, in the main application) with less
responsibilities.

Against it, when using gems like has_scope and pundit, we need to
explicitly specify which constant we need to be used on instantianting
an object: Project or Neighborly::Api::Project (since they don't have
access to Ruby's scopes of constants by modules).